+++ src/main.rs
@@ -1,3 +1,6 @@
+mod board;
+use board::Board;
+
 fn main() {
     let path = "example.csv".to_owned();
     let board = load_board(path);
@@ -7,7 +10,7 @@ fn main() {
 //this function takes a file system path, and returns a 2d vec of strings if it's a properly formatted
 //CSV, using the aptly named CSV crate. This is extremely volitile, and cannot handle improperly
 //formatted files
-fn load_board(path: String) -> Board {
+pub fn load_board(path: String) -> Board {
     let file = std::fs::File::open(path).unwrap();
 
     //let "has headers" be a command line option
@@ -20,48 +23,3 @@ fn load_board(path: String) -> Board {
     Board::new(board)
 }
 
-//Num for when a tile has a single state, to allow for faster iteration
-//Non for when a tile's entropy is not zero
-//consider in future using box and making it recursive
-enum Tile {
-    Num(u8),
-    Non(Vec<u8>)
-}
-
-//a simple 2d data structure to hold the sudoku board
-struct Board {
-    items: Vec<Vec<Tile>>
-}
-
-impl Board {
-    fn new(items: Vec<Vec<String>>) -> Board {
-        let mut board = vec![];
-        for i in items {
-            let mut row = vec![];
-            for j in i {
-                match j.parse::<u8>() {
-                    Ok(x) => row.push(Tile::Num(x)),
-                    //1, 2, 3 is placeholder
-                    Err(_) => row.push(Tile::Non(vec![1, 2, 3]))
-                }
-            }
-            board.push(row);
-        }
-        Board {
-            items: board
-        }
-    }
-
-    //iterates through a reference to the board's tiles, printing them out
-    fn display(&self) {
-        for i in &self.items {
-            for j in i {
-                match j {
-                    Tile::Num(x) => print!("{x}, "),
-                    Tile::Non(x) => print!("{:?}, ", x)
-                }
-            }
-            println!("");
-        }
-    }
-}
